<Code Optimization & Upgrade>
New:
- A new Dialog to display the full license content. Update the 'About'
  message as well.

Change:
- Remove <time.h> and use an alternative Win32 API to retrieve the
  current time.
- Reintroduce <strsafe.h> to handle strings.
- Refactor some functions and update most comments.
- Enable the precompiled header in this project.
- Update README.

+++ src/gamecore.h
@@ -1,13 +1,15 @@
 /*****************************************************************************\
  *  My Minesweepper -- a classic minesweeper game
- *  Copyright (C) 2020-2023 Gee W.
+ *  Copyright (C) 2020-2023 Gee Wang
  *
- *  This program is free software: you can redistribute it and/or modify
+ *  This file is part of My Minesweeper.
+ *
+ *  My Minesweeper is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
  *
- *  This program is distributed in the hope that it will be useful,
+ *  My Minesweeper is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *  GNU General Public License for more details.
@@ -29,52 +31,52 @@
 
 #pragma once
 
-#include "stdincludes.h"
+#include "stdafx.h"
 
 
 
 //Game Modes
-#define JUNIOR			0
-#define MIDDLE			1
-#define SENIOR			2
-#define CUSTOM			3
-#define CRUSH			4
-#define ISSTDMOD(Mode)	((BYTE)(Mode) < CUSTOM)
-#define ISCRUSH(Mode)	((BYTE)(Mode) >= CRUSH)
+#define JUNIOR          0
+#define MIDDLE          1
+#define SENIOR          2
+#define CUSTOM          3
+#define CRUSH           4
+#define ISSTDMOD(Mode)  ((BYTE)(Mode) < CUSTOM)
+#define ISCRUSH(Mode)   ((BYTE)(Mode) >= CRUSH)
 
-#define JUNIOR_WIDTH	9
-#define JUNIOR_HEIGHT	9
-#define JUNIOR_SIZE		(JUNIOR_WIDTH * JUNIOR_HEIGHT)
-#define JUNIOR_MINES	10
+#define JUNIOR_WIDTH    9
+#define JUNIOR_HEIGHT   9
+#define JUNIOR_SIZE     (JUNIOR_WIDTH * JUNIOR_HEIGHT)
+#define JUNIOR_MINES    10
 
-#define MIDDLE_WIDTH	16
-#define MIDDLE_HEIGHT	16
-#define MIDDLE_SIZE		(MIDDLE_WIDTH * MIDDLE_HEIGHT)
-#define MIDDLE_MINES	40
+#define MIDDLE_WIDTH    16
+#define MIDDLE_HEIGHT   16
+#define MIDDLE_SIZE     (MIDDLE_WIDTH * MIDDLE_HEIGHT)
+#define MIDDLE_MINES    40
 
-#define SENIOR_WIDTH	30
-#define SENIOR_HEIGHT	16
-#define SENIOR_SIZE		(SENIOR_WIDTH * SENIOR_HEIGHT)
-#define SENIOR_MINES	99
+#define SENIOR_WIDTH    30
+#define SENIOR_HEIGHT   16
+#define SENIOR_SIZE     (SENIOR_WIDTH * SENIOR_HEIGHT)
+#define SENIOR_MINES    99
 
-#define MAX_WIDTH		30
-#define MAX_HEIGHT		24
-#define MAX_SIZE		(MAX_WIDTH * MAX_HEIGHT)
-#define MIN_WIDTH		9
-#define MIN_HEIGHT		9
-#define MIN_SIZE		(MIN_WIDTH * MIN_HEIGHT)
-#define MAX_MINES(Size)	((WORD)((WORD)(Size) * 9 / 10))
-#define MIN_MINES		10
+#define MAX_WIDTH       30
+#define MAX_HEIGHT      24
+#define MAX_SIZE        (MAX_WIDTH * MAX_HEIGHT)
+#define MIN_WIDTH       9
+#define MIN_HEIGHT      9
+#define MIN_SIZE        (MIN_WIDTH * MIN_HEIGHT)
+#define MAX_MINES(Size) ((WORD)((WORD)(Size) * 9 / 10))
+#define MIN_MINES       10
 //end Game Modes
 
 //Game States
-#define INIT				0
-#define RUNNING				1
-#define SUCCESS				2
-#define FAIL				3
-#define UNKNOW				4
-#define ISBADSTATE(State)	((BYTE)(State) >= UNKNOW)
-#define ISGAMESET(State)	((BYTE)(State) > RUNNING)
+#define INIT                0
+#define RUNNING             1
+#define SUCCESS             2
+#define FAIL                3
+#define UNKNOW              4
+#define ISBADSTATE(State)   ((BYTE)(State) >= UNKNOW)
+#define ISGAMESET(State)    ((BYTE)(State) > RUNNING)
 //end Game States
 
 //Game Map Unit
@@ -83,56 +85,56 @@
  * |M|  S  |   N   |
  * +-+-+-+-+-+-+-+-+
  *     map_unit
- * M:	if this unit is a mine
- * S:	state of this unit:[covered, marked, flagged, uncovered, bombed, wrong flag]
- * N:	contains the number of mines in neighbors, from 0x0 to 0x8
+ * M:   If this unit is a mine.
+ * S:   State of this unit:[covered, marked, flagged, uncovered, bombed, wrong flag].
+ * N:   Contains the number of mines in neighbors, from 0x0 to 0x8.
  */
 
-#define MU_MINE			0x80
-#define MUS_COVER		0x00
-#define MUS_MARK		0x10
-#define MUS_FLAG		0x20
-#define MUS_UNCOV		0x30
-#define MUS_BOMB		0x40
-#define MUS_WRONG		0x50
-#define GETMUMINES(unit)		((BYTE)((unit) & 0x0F))
-#define GETMUSTATE(unit)		((BYTE)((unit) & 0x70))
-#define SETMUMINES(m, unit)		(unit = ((BYTE)((BYTE)((m) & 0x0F) | (BYTE)((unit) & 0xF0))))
-#define SETMUSTATE(S, unit)		(unit = ((BYTE)((BYTE)((S) & 0x70) | (BYTE)((unit) & 0x8F))))
-#define MUISMINE(unit)			((unit) & MU_MINE)
-#define MUISCLICKABLE(unit)		(GETMUSTATE(unit) <= MUS_MARK)
+#define MU_MINE         0x80
+#define MUS_COVER       0x00
+#define MUS_MARK        0x10
+#define MUS_FLAG        0x20
+#define MUS_UNCOV       0x30
+#define MUS_BOMB        0x40
+#define MUS_WRONG       0x50
+#define GETMUMINES(unit)        ((BYTE)((unit) & 0x0F))
+#define GETMUSTATE(unit)        ((BYTE)((unit) & 0x70))
+#define SETMUMINES(m, unit)     (unit = ((BYTE)((BYTE)((m) & 0x0F) | (BYTE)((unit) & 0xF0))))
+#define SETMUSTATE(S, unit)     (unit = ((BYTE)((BYTE)((S) & 0x70) | (BYTE)((unit) & 0x8F))))
+#define MUISMINE(unit)          ((unit) & MU_MINE)
+#define MUISCLICKABLE(unit)     (GETMUSTATE(unit) <= MUS_MARK)
 //end Game Map Unit
 
 //Game Score
-#define SCORE_NAME_LEN		20
-#define DEF_SCORE_NAME_EN	"anonymous"
+#define SCORE_NAME_LEN      20
+#define DEF_SCORE_NAME_EN   "anonymous"
 
-#define MAX_TIME	999
+#define MAX_TIME    999
 //end Game Score
 
 
 
 typedef struct _GameInfo {
-	BYTE mode;			//[junior, middle, senior, custom]
-	BYTE state;			//[init, running, success, fail]
-	bool mark;			//if the QuestionMark is used
-	BYTE width;			//GameMap width: map_units per line
-	BYTE height;		//GameMap height: map_units per column
-	WORD size;			//GameMap size: width x height
-	WORD mines;			//counts of mines in GameMap
-	short mine_remains;	//counts of mines that haven't been flagged
-	WORD uncov_units;	//counts of map_units that have been uncovered
-	WORD time;			//GameTime
-	BYTE map[MAX_SIZE];	//GameMap data, avaliable area depends on GameMode
+    BYTE mode;          //[junior, middle, senior, custom]
+    BYTE state;         //[init, running, success, fail]
+    bool mark;          //if the QuestionMark is used
+    BYTE width;         //GameMap width: map_units per line
+    BYTE height;        //GameMap height: map_units per column
+    WORD size;          //GameMap size: width x height
+    WORD mines;         //counts of mines in GameMap
+    short mine_remains; //counts of mines that haven't been flagged
+    WORD uncov_units;   //counts of map_units that have been uncovered
+    WORD time;          //GameTime
+    BYTE map[MAX_SIZE]; //GameMap data, avaliable area depends on GameMode
 } GameInfo, * PGameInfo;
 
 typedef struct _GameScore {
-	WORD junior_time;
-	WORD middle_time;
-	WORD senior_time;
-	TCHAR junior_name[SCORE_NAME_LEN];
-	TCHAR middle_name[SCORE_NAME_LEN];
-	TCHAR senior_name[SCORE_NAME_LEN];
+    WORD junior_time;
+    WORD middle_time;
+    WORD senior_time;
+    TCHAR junior_name[SCORE_NAME_LEN];
+    TCHAR middle_name[SCORE_NAME_LEN];
+    TCHAR senior_name[SCORE_NAME_LEN];
 } GameScore, * PGameScore;
 
 /* order in 2D:
@@ -143,7 +145,7 @@ typedef struct _GameScore {
  * +---+---+---+      +---+---+---+---+---+---+---+---+---+
  * | 6 | 7 | 8 |
  * +---+---+---+
- * stores index of each neighbor, -1 means no such neighbor
+ * Stores index of each neighbor, -1 means no such neighbor.
  */
 typedef int Neighbor[9];
 
@@ -156,7 +158,7 @@ typedef int Neighbor[9];
  *      ----->
  *       0   1   2   3
  *     +---+---+---+---+
- * y| 0| 0 | 1 | 2 | 3 |	eg: index = 2  <==>  (x, y) = (2, 0)
+ * y| 0| 0 | 1 | 2 | 3 |    eg: index = 2  <==>  (x, y) = (2, 0)
  *  |  +---+---+---+---+
  *  v 1| 4 |...index...|
  *     +---+---+---+---+
@@ -267,7 +269,7 @@ void resetRecord(PGameScore pScore);
 
 //Get the best time under given GameMode.
 //Return -1 if error.
-DWORD getRecordTime(PGameScore pScore, BYTE gamemode);
+WORD getRecordTime(PGameScore pScore, BYTE gamemode);
 
 //Get the player name under given GameMode.
 //This function returns the head address of TCHAR[SCORE_NAME_LEN],
